+++ navground_sim_py/navground/sim/ui/to_svg.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-import inspect
+import dataclasses as dc
 import math
 import os
 from collections import ChainMap
-from collections.abc import Callable, Collection, MutableMapping
-from typing import TYPE_CHECKING, Any, TypeAlias, cast
+from collections.abc import Callable, Collection
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 import numpy.typing
@@ -13,38 +13,10 @@ from navground import core
 
 if TYPE_CHECKING:
     from .. import (Agent, Entity, Obstacle, Wall, World)
+    from .common import Attributes, Point, Decorate, Rect
 
-from ..bounds import Bounds, bounds_for_world
-
-Rect: TypeAlias = 'Bounds | np.typing.NDArray[np.floating[Any]]'
-Attributes: TypeAlias = MutableMapping[str, str]
-Point: TypeAlias = core.Vector2
-EntityDecorator: TypeAlias = 'Callable[[Entity], Attributes]'
-WorldDecorator: TypeAlias = 'Callable[[Entity, World], Attributes]'
-Decorate: TypeAlias = 'EntityDecorator | WorldDecorator'
-
-
-def wrap_as_world_decorator(decorate: Decorate) -> WorldDecorator:
-    sig = inspect.signature(decorate)
-    if len(sig.parameters) == 1:
-
-        def f(e: Entity, w: World) -> Attributes:
-            return cast(EntityDecorator, decorate)(e)
-
-        return f
-    return cast(WorldDecorator, decorate)
-
-
-def rect_around(center: core.Vector2, width: float | None,
-                height: float | None) -> Rect:
-    if width is None:
-        width = height
-    if height is None:
-        height = width
-    if height is None or width is None:
-        raise ValueError("Provide at least one of height or width")
-    delta = np.array((width * 0.5, height * 0.5))
-    return center - delta, center + delta
+from ..bounds import bounds_for_world
+from .common import render_default_config, wrap_as_world_decorator
 
 
 def svg_color(r: float, g: float, b: float) -> str:
@@ -202,100 +174,61 @@ def svg_for_agent(
                      delta=delta)
 
 
-UNSPECIFIED: Any = object()
-
-
-def svg_for_world(
-    world: World | None = None,
-    *,
-    precision: int = UNSPECIFIED,
-    decorate: Decorate | None = UNSPECIFIED,
-    bounds: Rect | None = UNSPECIFIED,
-    width: int = UNSPECIFIED,
-    min_height: int = UNSPECIFIED,
-    relative_margin: float = UNSPECIFIED,
-    background_color: str = UNSPECIFIED,
-    display_shape: bool = UNSPECIFIED,
-    display_safety_margin: bool = UNSPECIFIED,
-    grid: float = UNSPECIFIED,
-    grid_color: str = UNSPECIFIED,
-    grid_thickness: float = UNSPECIFIED,
-    rotation: tuple[core.Vector2, float] | float | None = UNSPECIFIED,
-    extras: Collection[Callable[[World], str]] = UNSPECIFIED,
-    background_extras: Collection[Callable[[World],
-                                           str]] = UNSPECIFIED) -> str:
+def svg_for_world(world: World | None = None, **kwargs: Any) -> str:
     """
     Draw the world as a SVG.
 
-    :param      world:                  The world to display
-    :param      precision:              The number of decimal digits for poses
-    :param      decorate:               A function to decorate entities.
-                                        Should return a dictionary of valid SVG style attributes,
-                                        e.g., ``{"fill": "red"}`` for a given entity.
-    :param      bounds:                 The rectangular area to be displayed
-    :param      width:                  The width in pixels
-    :param      min_height:             The minimum height in pixels
-    :param      relative_margin:        The relative margin
-    :param      background_color:       A valid SVG color for the background
-    :param      display_shape:          Whether to display the agent circular shape
-    :param      display_safety_margin:  Whether to display the agent safety margin
-    :param      grid:                   The size of the square grid tile
-                                        (set to zero or negative to skip drawing a grid)
-    :param      grid_color:             The color of the grid
-    :param      grid_thickness:         The thickness of the grid
-    :param      rotation:               A planar rotation applied before drawing [rad]
-    :param      extras:                 Provides extras rendering added at the end of to the svg
-    :param      background_extras:      Provides extras rendering added at the
-                                        beginning of to the svg
+    :param      world:   The world to display
+    :param      kwargs:  Optional configuration:
+        same fields as :py:class:`navground.sim.ui.RenderConfig`.
+
+    The actual configuration is computed by looking (in order) to
+
+    1. the arguments of this function;
+    2. the world-specific configuration :py:attr:`navground.sim.World.render_kwargs`;
+    3. the default configuration :py:attr:`navground.sim.ui.render_default_config`.
 
     :returns:   An SVG string
     """
-    return svg_for_world_and_dims(**locals())[0]
+    return svg_for_world_and_dims(world, **kwargs)[0]
 
 
 def svg_for_world_and_dims(world: World | None = None,
                            **kwargs: Any) -> tuple[str, dict[str, str]]:
-    spec = inspect.getfullargspec(_svg_for_world_and_dims)
+    full_kwargs = dc.asdict(render_default_config)
     if world:
-        skwargs = {
-            k: v
-            for k, v in world.render_kwargs.items() if k in spec.kwonlyargs
-        }
-    else:
-        skwargs = {}
-    skwargs.update({
-        k: v
-        for k, v in kwargs.items()
-        if v is not UNSPECIFIED and k in spec.kwonlyargs
-    })
-    return _svg_for_world_and_dims(world=world, **skwargs)
+        full_kwargs.update(world.render_kwargs)
+    full_kwargs.update(kwargs)
+    return _svg_for_world_and_dims(world=world, **full_kwargs)
 
 
-def _svg_for_world_and_dims(
-    world: World | None = None,
-    *,
-    prefix: str = '',
-    precision: int = 2,
-    decorate: Decorate | None = None,
-    standalone: bool = True,
-    bounds: Rect | None = None,
-    width: int = 600,
-    min_height: int = 100,
-    relative_margin: float = 0.05,
-    include_default_style: bool = True,
-    external_style_path: str = '',
-    style: str = '',
-    background_color: str = 'snow',
-    display_shape: bool = False,
-    display_safety_margin: bool = False,
-    grid: float = 0,
-    grid_color: str = 'grey',
-    grid_thickness: float = 0.01,
-    rotation: tuple[core.Vector2, float] | float | None = None,
-    extras: Collection[Callable[[World], str]] = [],
-    background_extras: Collection[Callable[[World], str]] = []
-) -> tuple[str, dict[str, str]]:
+def _svg_for_world_and_dims(world: World | None = None,
+                            *,
+                            prefix: str = '',
+                            precision: int = 2,
+                            decorate: Decorate | None = None,
+                            standalone: bool = True,
+                            bounds: Rect | None = None,
+                            width: int = 600,
+                            min_height: int = 100,
+                            relative_margin: float = 0.05,
+                            include_default_style: bool = True,
+                            external_style_path: str = '',
+                            style: str = '',
+                            background_color: str = 'snow',
+                            display_shape: bool = False,
+                            display_safety_margin: bool = False,
+                            grid: float = 0,
+                            grid_color: str = 'grey',
+                            grid_thickness: float = 0.01,
+                            rotation: tuple[core.Vector2, float] | float
+                            | None = None,
+                            extras: Collection[Callable[[World], str]] = [],
+                            background_extras: Collection[Callable[[World],
+                                                                   str]] = [],
+                            **kwargs: Any) -> tuple[str, dict[str, str]]:
     import jinja2
+
     from .. import BoundingBox
 
     g = ""
@@ -331,26 +264,6 @@ def _svg_for_world_and_dims(
                     display_shape,
                     display_safety_margin,
                     delta=delta)
-
-        # g = "<g id='_world'>"
-        # for wall in world.walls:
-        #     g += svg_for_wall(wall, precision, prefix,
-        #                       decorate(wall) if decorate else {})
-        # for obstacle in world.obstacles:
-        #     g += svg_for_obstacle(obstacle, precision, prefix,
-        #                           decorate(obstacle) if decorate else {})
-        # for agent in world.agents:
-        #     g += svg_for_agent(agent, precision, prefix,
-        #                        decorate(agent) if decorate else {},
-        #                        display_shape)
-        # if bounds is None:
-        #     bb = world.bounding_box
-        #     bounds = bb.p1, bb.p2
-        # width, height, view_box, min_y = size(bounds, width, min_height,
-        #                                       relative_margin)
-        # g += "</g>"
-        # for delta in world.get_lattice_grid(include_zero=False, c8=True):
-        #     g += f'<use href="#_world" x="{delta[0]}" y="{delta[1]}" />'
     else:
         height = width
         view_box = (0, 0, 1, 1)
@@ -415,25 +328,6 @@ def _svg_for_world_and_dims(
         epilog=epilog,
         **grid_kwargs,
         **dims), dims
-
-
-# def world_bound(world: World) -> Rect:
-#     ps = []
-#     for wall in world.walls:
-#         ps.append(wall.line.p1)
-#         ps.append(wall.line.p2)
-#     for obstacle in world.obstacles:
-#         delta = np.ones(2) * obstacle.disc.radius
-#         p = obstacle.disc.position
-#         ps.append(p - delta)
-#         ps.append(p + delta)
-#     for agent in world.agents:
-#         delta = np.ones(2) * agent.radius
-#         ps.append(agent.position - delta)
-#         ps.append(agent.position + delta)
-#     if not ps:
-#         return np.zeros(2), np.zeros(2)
-#     return np.min(ps, axis=0), np.max(ps, axis=0)
 
 
 def size(